requirements/dev.txt
-r base.txt

black==24.10.0
flake8==7.1.1
isort==5.13.2
flake8-django==1.4
flake8-bugbear==24.10.31
pylint==3.3.2
pylint-django==2.6.1
pre-commit==4.0.1
django-debug-toolbar==4.4.6
factory-boy==3.3.1
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
